+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Improved CalDAV discovery - [#363](https://github.com/hkdb/aerion/issues/363)
 - Basic CarDAV (Mailfence) support - [#366](https://github.com/hkdb/aerion/issues/366)
 - Fixed typo in autostart code - [#33](https://github.com/hkdb/aerion/issues/33)
+- Updated nb translations - [PR #371](https://github.com/hkdb/aerion/pull/371)
 
 
 **v0.3.2 - 07-16-2026**